Fix text-style rule.
For the text style, we can try to catch a 'none' mixed with other things, and that is why we say that at the top level, a text style is either 'none' or the recursive rule with multiple styles.
The recursive rule, in turn, will accept any of the known values or variables, in any non-zero number.