Add selector rule.
The selector rule works by just looking for something that looks like the initial character of a selector (in Textual, that's not that many characters) and then the end of the selector is when the rule list actually starts.
Then, another rule (selector-innards) is used to make sure the selector makes sense.