Finish adding selector-innards.
We simplify the rule that looks for invalid selectors.
We skip the rule that looks for things that look like [...] because we don't have that in TCSS.
We skip pseudo-elements and functional pseudo-classes.
We modify the rule about custom HTML tags and use it for all the widget class names that TCSS uses.